Refactor S3 config utils to hdds-framework
